Added applications fragment
- Changed the layout of the applications fragment with the latest figma designs.
- Added a search bar on the top, and fixed issue of applications loading every tab change.
- Also added click listeners to the applications item.

+++ app/src/main/java/com/studbudd/application_tracker/MainActivity.kt
@@ -169,8 +169,8 @@ class MainActivity : AppCompatActivity() {
     override fun onNewIntent(intent: Intent?) {
         super.onNewIntent(intent)
         if (intent != null) {
-            val applicationId: Int = intent.getIntExtra(NotifyWorker.applicationIdKey, -1)
-            if (applicationId != -1) {
+            val applicationId: Long = intent.getLongExtra(NotifyWorker.applicationIdKey, -1)
+            if (applicationId != -1L) {
                 navController.navigate(
                     ApplicationsFragmentDirections.actionApplicationsFragmentToApplicationDetailsFragment(
                         applicationId
